HCMPRE-2428: Refactor App Configuration Redesign
