+++ health/micro-ui/web/micro-ui-internals/packages/modules/campaign-manager/src/pages/employee/appConfigurationRedesign/AppConfigurationParentLayer.js
@@ -1,28 +1,11 @@
+import React, { useEffect, useMemo, useReducer, useState } from "react";
 import { Loader, Stepper, Toast, Tooltip } from "@egovernments/digit-ui-components";
 import { Header } from "@egovernments/digit-ui-react-components";
-import React, { useEffect, useMemo, useReducer, useState } from "react";
 import { useTranslation } from "react-i18next";
-import ImpelComponentWrapper from "./ImpelComponentWrapper";
 import { useHistory } from "react-router-dom";
-// import { dummyMaster } from "../../../configs/dummyMaster";
-
-const Tabs = ({ numberTabs, onTabChange }) => {
-  const { t } = useTranslation();
-  return (
-    <div className="configure-app-tabs">
-      {numberTabs.map((_, index) => (
-        <button
-          key={index}
-          type="button"
-          className={`configure-app-tab-head ${_.active === true ? "active" : ""} hover`}
-          onClick={() => onTabChange(_, index)}
-        >
-          <p style={{ margin: 0, position: "relative", top: "-0 .1rem" }}>{t(_.parent)}</p>
-        </button>
-      ))}
-    </div>
-  );
-};
+import ImpelComponentWrapper from "./ImpelComponentWrapper";
+import { restructure, reverseRestructure } from "../../../utils/appConfigHelpers";
+import Tabs from "./Tabs";
 
 const dispatcher = (state, action) => {
   switch (action.key) {
@@ -41,7 +24,6 @@ const dispatcher = (state, action) => {
           }
           return i;
         }),
-        // appData: state.appData ? [...state.appData, ...action.data] : [...action.data],
       };
     case "SETFORM":
       return {
@@ -53,186 +35,7 @@ const dispatcher = (state, action) => {
   }
 };
 
-// Updated getTypeAndMetaData to use appfieldtype master data from MDMS
-const getTypeAndMetaData = (field, fieldTypeMasterData = []) => {
-  if (!Array.isArray(fieldTypeMasterData) || fieldTypeMasterData.length === 0) {
-    return { type: "textInput" };
-  }
-
-  // Try to find a matching field type from master data
-  const matched = fieldTypeMasterData.find((item) => {
-    // Match both type and format from metadata
-    return item?.metadata?.type === field.type && item?.metadata?.format === field.format;
-  });
-
-  if (!matched) {
-    return { type: "textInput" };
-  }
-
-  // Start with the fieldType as type
-  let result = { type: matched.fieldType, appType: matched.type };
-
-  // Copy all metadata properties except type/format (already used)
-  Object.entries(matched.metadata || {}).forEach(([key, value]) => {
-    if (key !== "type" && key !== "format") {
-      result[key] = value;
-    }
-  });
-
-  // Map/rename additional attributes as per attributeToRename
-  if (matched.attributeToRename) {
-    Object.entries(matched.attributeToRename).forEach(([to, from]) => {
-      if (field[from] !== undefined) {
-        result[to] = field[from];
-      }
-    });
-  }
-
-  // Special handling for enums to dropdownOptions (for dropdown/select)
-  if (matched.fieldType === "dropdown" && field.enums) {
-    result.dropDownOptions = [...field.enums];
-  }
-
-  // Pass through other common field properties if needed (e.g., min, max, prefix)
-  ["min", "max", "prefix", "startDate", "endDate"].forEach((prop) => {
-    if (field[prop] !== undefined) {
-      result[prop] = field[prop];
-    }
-  });
-
-  return result;
-};
-
-const restructure = (data1, fieldTypeMasterData = [], parent) => {
-  return data1
-    ?.sort((a, b) => a.order - b.order)
-    .map((page) => {
-      const cardFields = page.properties
-        ?.sort((a, b) => a.order - b.order)
-        ?.map((field, index) => ({
-          label: field.label || "",
-          value: field.value || "",
-          active: true,
-          jsonPath: field.fieldName || "",
-          metaData: {},
-          Mandatory: field.required || false,
-          hidden: field.hidden || false,
-          deleteFlag: field.deleteFlag || false,
-          isLocalised: field.isLocalised ? true : false,
-          innerLabel: field.innerLabel || "",
-          helpText: field.helpText || "",
-          errorMessage: field?.errorMessage || "",
-          tooltip: field.tooltip || "",
-          infoText: field.infoText || "",
-          order: field.order,
-          ...getTypeAndMetaData(field, fieldTypeMasterData),
-        }));
-
-      return {
-        name: page.label || page.page || "UNKNOWN",
-        cards: [
-          {
-            header: crypto.randomUUID(),
-            fields: cardFields,
-            headerFields: [
-              {
-                type: "text",
-                label: "SCREEN_HEADING",
-                value: page.label || "",
-                active: true,
-                jsonPath: "ScreenHeading",
-                metaData: {},
-                required: true,
-                isLocalised: page.label ? true : false,
-              },
-              {
-                type: "textarea",
-                label: "SCREEN_DESCRIPTION",
-                value: page.description || "",
-                active: true,
-                jsonPath: "Description",
-                metaData: {},
-                required: true,
-                isLocalised: page.description ? true : false,
-              },
-            ],
-          },
-        ],
-        actionLabel: page?.actionLabel || "",
-        order: page.order,
-        config: {
-          enableComment: false,
-          enableFieldAddition: true,
-          allowFieldsAdditionAt: ["body"],
-          enableSectionAddition: false,
-          allowCommentsAdditionAt: ["body"],
-        },
-        parent: parent?.name || "",
-      };
-    });
-};
-
-const guessPageName = (label) => {
-  const map = {
-    BENE_LOCATION: "beneficiaryLocation",
-    BENE_HOUSE: "HouseDetails",
-    // Add more mappings as needed
-  };
-  return map[label] || label;
-};
-
-// Helper to get type/format and handle attribute renaming from appType using fieldTypeMasterData
-const getTypeAndFormatFromAppType = (field, fieldTypeMasterData = []) => {
-  if (!field.appType) return {};
-  const matched = fieldTypeMasterData.find((item) => item.type === field.appType);
-  if (!matched) return {};
-  const result = {
-    type: matched.metadata?.type,
-    format: matched.metadata?.format,
-  };
-  // Handle attributeToRename: { targetKey: sourceKey }
-  if (matched.attributeToRename) {
-    Object.entries(matched.attributeToRename).forEach(([targetKey, sourceKey]) => {
-      result[sourceKey] = field[targetKey];
-    });
-  }
-  return result;
-};
-
-// Update reverseRestructure to use getTypeAndFormatFromAppType
-const reverseRestructure = (updatedData, fieldTypeMasterData = []) => {
-  return updatedData.map((section, index) => {
-    const properties = section.cards?.[0]?.fields.map((field, fieldIndex) => {
-      const typeAndFormat = getTypeAndFormatFromAppType(field, fieldTypeMasterData);
-      return {
-        label: field.label || "",
-        order: fieldIndex + 1,
-        value: field.value || "",
-        hidden: false, // can't be derived from updatedData unless explicitly added
-        required: field.Mandatory || false,
-        hidden: field.hidden || false,
-        fieldName: field.jsonPath || "",
-        helpText: field.helpText || "",
-        tooltip: field.tooltip || "",
-        infoText: field.infoText || "",
-        innerLabel: field.innerLabel || "",
-        errorMessage: field.errorMessage || "",
-        deleteFlag: field.deleteFlag || false,
-        ...typeAndFormat,
-      };
-    });
-
-    return {
-      page: guessPageName(section.name),
-      type: "object",
-      label: section.cards?.[0]?.headerFields?.find((i) => i.jsonPath === "ScreenHeading")?.value,
-      description: section.cards?.[0]?.headerFields?.find((i) => i.jsonPath === "Description")?.value,
-      actionLabel: section?.actionLabel || "",
-      order: index + 1,
-      properties,
-    };
-  });
-};
+const mdms_context_path = window?.globalConfigs?.getConfig("MDMS_V2_CONTEXT_PATH") || "mdms-v2";
 
 const AppConfigurationParentRedesign = () => {
   const tenantId = Digit.ULBService.getCurrentTenantId();
@@ -257,6 +60,7 @@ const AppConfigurationParentRedesign = () => {
     }
     return null;
   }, [parentState?.actualTemplate?.name, parentState?.actualTemplate?.project]);
+
   const { isLoading: isLoadingAppConfigMdmsData, data: AppConfigMdmsData } = Digit.Hooks.useCustomMDMS(
     Digit.ULBService.getCurrentTenantId(),
     MODULE_CONSTANTS,
@@ -277,7 +81,6 @@ const AppConfigurationParentRedesign = () => {
     { schemaCode: "BASE_APP_MASTER_DATA3" } //mdmsv2
   );
 
-  const mdms_context_path = window?.globalConfigs?.getConfig("MDMS_V2_CONTEXT_PATH") || "mdms-v2";
   const reqCriteriaForm = {
     url: `/${mdms_context_path}/v2/_search`,
     body: {
@@ -297,20 +100,8 @@ const AppConfigurationParentRedesign = () => {
     },
   };
 
-  const correctField = (key) => {
-    switch (key) {
-      case "text":
-        return "textInput";
-      case "date":
-        return "datePicker";
-      default:
-        return key;
-    }
-  };
-
   const { isLoading, data: formData } = Digit.Hooks.useCustomAPIHook(reqCriteriaForm);
 
-  // const { mutate: updateMutate } = Digit.Hooks.campaign.useUpdateFormBuilderConfig(tenantId);
   const { mutate: updateMutate } = Digit.Hooks.campaign.useUpdateAppConfig(tenantId);
 
   useEffect(() => {
@@ -318,12 +109,6 @@ const AppConfigurationParentRedesign = () => {
       setTimeout(closeToast, 10000);
     }
   }, [showToast]);
-
-  // useEffect(() => {
-  //   if (!NewisLoadingAppConfigMdmsData && NewAppConfigMdmsData) {
-  //     const temp = restructure(NewAppConfigMdmsData?.SimplifiedAppConfig?.[0]?.pages);
-  //   }
-  // }, [NewAppConfigMdmsData, NewisLoadingAppConfigMdmsData]);
 
   useEffect(() => {
     if (!isLoading && formData && formId && AppConfigMdmsData?.[fieldTypeMaster]?.length > 0) {
@@ -336,14 +121,6 @@ const AppConfigurationParentRedesign = () => {
         appIdData: formData?.data,
       });
     }
-    // } else if (!isLoadingAppConfigMdmsData && AppConfigMdmsData?.[masterName]) {
-    //   const temp = restructure(AppConfigMdmsData?.[masterName]?.[0]?.pages);
-    //   parentDispatch({
-    //     key: "SET",
-    //     data: [...temp],
-    //     template: AppConfigMdmsData?.[masterName],
-    //   });
-    // }
   }, [isLoadingAppConfigMdmsData, AppConfigMdmsData, formData]);
 
   useEffect(() => {
@@ -388,6 +165,11 @@ const AppConfigurationParentRedesign = () => {
       isSubmit: stepper?.find((i) => i.active)?.isLast ? true : false,
     });
     if (stepper?.find((i) => i.active)?.isLast) {
+      debugger;
+      const mergedTemplate = parentState.currentTemplate.map((item) => {
+        const updated = screenData.find((d) => d.name === item.name);
+        return updated ? updated : item;
+      });
       const reverseData = reverseRestructure(parentState?.currentTemplate, AppConfigMdmsData?.[fieldTypeMaster]);
       // const nextTabAvailable = numberTabs.some((tab) => tab.code > currentStep.code && tab.active);
       const reverseFormat = {
@@ -396,10 +178,7 @@ const AppConfigurationParentRedesign = () => {
         pages: reverseData,
       };
 
-      const updatedFormData = formData;
-
-      // Update the data property while maintaining the required structure
-      updatedFormData.data = reverseFormat;
+      const updatedFormData = { ...formData, data: reverseFormat };
 
       await updateMutate(
         {
@@ -441,9 +220,11 @@ const AppConfigurationParentRedesign = () => {
       setCurrentStep((prev) => prev + 1);
     }
   };
+
   const closeToast = () => {
     setShowToast(null);
   };
+
   const back = () => {
     if (stepper?.find((i) => i.active)?.isFirst) {
       setShowToast({ key: "error", label: "CANNOT_GO_BACK" });
@@ -451,6 +232,7 @@ const AppConfigurationParentRedesign = () => {
       setCurrentStep((prev) => prev - 1);
     }
   };
+
   return (
     <div>
       <Header className="app-config-header">{t(`${currentScreen?.[0]?.name}`)}</Header>
